presentationからcontainerに処理を移行
